Modified migrate.sh so it can be run locally or piped in via curl

+++ migrate.sh
@@ -1,11 +1,21 @@
 #!/bin/bash -e
+
+if [ "${BASH_SOURCE[0]}" == "" ]; then
+	RUNNING_LOCALLY=false
+	TMPDIR=`mktemp -d /tmp/ios-starter.XXXXXX`	
+else
+	RUNNING_LOCALLY=true
+	TMPDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+fi
+
 echo
-echo -n "Application Name (e.g. 'MyApplication' with no spaces)? "
+echo -n "Application Name (e.g. 'MyApplication' with no spaces will be created in current dir)? "
 read PROJECT_NAME
 mkdir "$PROJECT_NAME"
 
-TMPDIR=`mktemp -d /tmp/ios-starter.XXXXXX`
-git clone git@github.com:twobitlabs/iOSXcodeStarterProject.git $TMPDIR
+if ! $RUNNING_LOCALLY; then
+	git clone git@github.com:twobitlabs/iOSXcodeStarterProject.git $TMPDIR
+fi
 cd "$PROJECT_NAME"
 
 cp -fr $TMPDIR/Tests \
@@ -40,7 +50,9 @@ git add .
 git commit -a -m "Initial Commit"
 git submodule update --init --recursive
 
-rm -rf $TMPDIR
+if ! $RUNNING_LOCALLY ; then
+	rm -rf $TMPDIR
+fi
 
 echo "Success! Launching $PROJECT_NAME"
 open $PROJECT_NAME.xcodeproj
